Do not increment rank for NA:
by default R's rank() puts the NAs at the end, with a rank higher than the previous. First version was copying this behavior, but that becomes a problem when reusing parameters for a validation set with NAs: we can end up with a rank higher than the number of rows in the reference matrix, hence infinity transformed values. New version keeps NAs at the ceiling highest rank from x.